+++ src/integTest/kotlin/snyk/oss/annotator/OSSGradleAnnotatorTest.kt
@@ -4,9 +4,7 @@ import com.google.gson.Gson
 import com.intellij.codeInsight.intention.IntentionAction
 import com.intellij.lang.annotation.AnnotationBuilder
 import com.intellij.lang.annotation.AnnotationHolder
-import com.intellij.openapi.application.WriteAction
 import com.intellij.openapi.util.TextRange
-import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiFile
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.testFramework.replaceService
@@ -17,6 +15,8 @@ import io.mockk.unmockkAll
 import io.mockk.verify
 import io.snyk.plugin.pluginSettings
 import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.plugins.groovy.GroovyFileType
+import org.jetbrains.plugins.groovy.lang.psi.GroovyFile
 import org.junit.Before
 import org.junit.Test
 import snyk.common.SnykCachedResults
@@ -27,20 +27,29 @@ import snyk.oss.Vulnerability
 import java.nio.file.Paths
 
 @Suppress("DuplicatedCode", "FunctionName")
-class OSSGradleKtsAnnotatorTest : BasePlatformTestCase() {
-    private var cut = OSSGradleKtsAnnotator()
+class OSSGradleAnnotatorTest : BasePlatformTestCase() {
+    private val cut by lazy { OSSGradleAnnotator() }
     private val annotationHolderMock = mockk<AnnotationHolder>(relaxed = true)
-    private val buildGradleKtsFilename = "build.gradle.kts"
-    private val ossResultGradleKts =
-        javaClass.classLoader.getResource("oss-test-results/oss-result-gradle-kts.json")!!.readText(Charsets.UTF_8)
 
-    private lateinit var virtualFileBuildGradleKts: VirtualFile
-    private lateinit var buildGradleKts: KtFile
+    private val ossResultGradleKtsJson =
+        javaClass.classLoader.getResource("oss-test-results/oss-result-gradle-kts.json")!!.readText(Charsets.UTF_8)
+    private val buildGradleKts: KtFile by lazy {
+        myFixture.configureByFile("build.gradle.kts")!! as KtFile
+    }
+    private val ossResultGradleJson =
+        javaClass.classLoader.getResource("oss-test-results/oss-result-gradle.json")!!.readText(Charsets.UTF_8)
+    private val buildGradle: GroovyFile by lazy {
+        val buildGradleText = OSSGradleAnnotator::class.java
+            .getResource("/test-fixtures/oss/annotator/build.gradle")!!.readText(Charsets.UTF_8)
+        val groovyFile = myFixture.configureByText(GroovyFileType.GROOVY_FILE_TYPE, buildGradleText)!! as GroovyFile
+        groovyFile.name = "build.gradle"
+        return@lazy groovyFile
+    }
 
     private val snykCachedResults: SnykCachedResults = mockk(relaxed = true)
 
     override fun getTestDataPath(): String {
-        val resource = OSSGradleKtsAnnotator::class.java.getResource("/test-fixtures/oss/annotator")
+        val resource = OSSGradleAnnotator::class.java.getResource("/test-fixtures/oss/annotator")
         requireNotNull(resource) { "Make sure that the resource $resource exists!" }
         return Paths.get(resource.toURI()).toString()
     }
@@ -53,11 +62,6 @@ class OSSGradleKtsAnnotatorTest : BasePlatformTestCase() {
         unmockkAll()
         project.replaceService(SnykCachedResults::class.java, snykCachedResults, project)
         pluginSettings().fileListenerEnabled = false
-        virtualFileBuildGradleKts = myFixture.copyFileToProject(buildGradleKtsFilename)
-        buildGradleKts = WriteAction.computeAndWait<KtFile, Throwable> {
-            psiManager.findFile(virtualFileBuildGradleKts)!! as KtFile
-        }
-        cut = OSSGradleKtsAnnotator()
     }
 
     override fun tearDown() {
@@ -77,11 +81,20 @@ class OSSGradleKtsAnnotatorTest : BasePlatformTestCase() {
     }
 
     @Test
+    fun `test gradle apply should trigger newAnnotation call`() {
+        every { snykCachedResults.currentOssResults } returns createGradleOssResultWithIssues()
+
+        cut.apply(buildGradle, Unit, annotationHolderMock)
+
+        verify(exactly = 6) { annotationHolderMock.newAnnotation(any(), any()) }
+    }
+
+    @Test
     fun `test textRange for Gradle Kts pom`() {
         val ossResult = createGradleKtsOssResultWithIssues()
         val vulnerabilities = ossResult.allCliIssues!![0].vulnerabilities
 
-        var issue = vulnerabilities[0] // org.jetbrains.kotlin:kotlin-test-junit")
+        var issue = vulnerabilities[0] // org.jetbrains.kotlin:kotlin-test-junit
         var expectedStart = 980
         var expectedEnd = 1018
         checkRangeFinding(expectedStart, expectedEnd, issue, buildGradleKts)
@@ -107,11 +120,60 @@ class OSSGradleKtsAnnotatorTest : BasePlatformTestCase() {
         checkRangeFinding(expectedStart, expectedEnd, issue, buildGradleKts)
     }
 
+    @Test
+    fun `test textRange for Gradle pom`() {
+        val ossResult = createGradleOssResultWithIssues()
+        val vulnerabilities = ossResult.allCliIssues!![0].vulnerabilities
+
+        // 'com.google.guava:guava:29.0-jre'
+        checkRangeFinding(
+            expectedStart = 720,
+            expectedEnd = 753,
+            issue = vulnerabilities[0],
+            psiFile = buildGradle
+        )
+        // 'junit:junit:4.13'
+        checkRangeFinding(
+            expectedStart = 874,
+            expectedEnd = 892,
+            issue = vulnerabilities[1],
+            psiFile = buildGradle
+        )
+        // 'org.apache.logging.log4j:log4j-core:2.14.1'
+        checkRangeFinding(
+            expectedStart = 773,
+            expectedEnd = 817,
+            issue = vulnerabilities[2],
+            psiFile = buildGradle
+        )
+        // 'org.apache.logging.log4j:log4j-core:2.14.1'
+        checkRangeFinding(
+            expectedStart = 773,
+            expectedEnd = 817,
+            issue = vulnerabilities[3],
+            psiFile = buildGradle
+        )
+        // 'org.apache.logging.log4j:log4j-core:2.14.1'
+        checkRangeFinding(
+            expectedStart = 773,
+            expectedEnd = 817,
+            issue = vulnerabilities[4],
+            psiFile = buildGradle
+        )
+        // 'org.apache.logging.log4j:log4j-core:2.14.1'
+        checkRangeFinding(
+            expectedStart = 773,
+            expectedEnd = 817,
+            issue = vulnerabilities[5],
+            psiFile = buildGradle
+        )
+    }
+
     private fun checkRangeFinding(
         expectedStart: Int,
         expectedEnd: Int,
         issue: Vulnerability,
-        psiFile: PsiFile = buildGradleKts
+        psiFile: PsiFile
     ) {
         val expectedRange = TextRange(expectedStart, expectedEnd)
         val actualRange = cut.textRange(psiFile, issue)
@@ -150,5 +212,8 @@ class OSSGradleKtsAnnotatorTest : BasePlatformTestCase() {
     }
 
     private fun createGradleKtsOssResultWithIssues(): OssResult =
-        OssResult(listOf(Gson().fromJson(ossResultGradleKts, OssVulnerabilitiesForFile::class.java)))
+        OssResult(listOf(Gson().fromJson(ossResultGradleKtsJson, OssVulnerabilitiesForFile::class.java)))
+
+    private fun createGradleOssResultWithIssues(): OssResult =
+        OssResult(listOf(Gson().fromJson(ossResultGradleJson, OssVulnerabilitiesForFile::class.java)))
 }