+++ src/main/kotlin/io/snyk/plugin/ui/toolwindow/NavigatableToSourceTreeNode.kt
@@ -1,0 +1,5 @@
+package io.snyk.plugin.ui.toolwindow
+
+interface NavigatableToSourceTreeNode {
+    val navigateToSource: () -> Unit
+}