+++ src/main/kotlin/snyk/iac/ui/toolwindow/IacIssueTreeNode.kt
@@ -1,10 +1,12 @@
 package snyk.iac.ui.toolwindow
 
 import com.intellij.openapi.project.Project
+import io.snyk.plugin.ui.toolwindow.NavigatableToSourceTreeNode
 import snyk.iac.IacIssue
 import javax.swing.tree.DefaultMutableTreeNode
 
 class IacIssueTreeNode(
     issue: IacIssue,
-    val project: Project
-) : DefaultMutableTreeNode(issue)
+    val project: Project,
+    override val navigateToSource: () -> Unit
+) : DefaultMutableTreeNode(issue), NavigatableToSourceTreeNode